Refactor: EditTeamPost컴포넌트에 WithLoading적용
