+++ webapp/src/pages/EditTeamPost/style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const Board = styled.div`
+export const Container = styled.div`
   display: flex;
   flex-direction: column;
   width: 600px;
@@ -10,8 +10,4 @@ export const Board = styled.div`
   border-radius: 15px;
   box-shadow: 2px 2px 2px 2px gray;
   flex-wrap: wrap;
-`;
-
-export const MdEditorContainer = styled.div`
-  width: 100%;
 `;
